feat: feature flag all my account pages (#2801)
## What's the purpose of this pull request?

This pull is to centralize and streamline the feature flag check and
redirection logic for my account feature.

## How it works?

* Added `getMyAccountRedirect` in `src/utils/myAccountRedirect.tsx`.
This function checks if the Faststore My Account feature is enabled and
determines the appropriate redirection destination based on the feature
flag and query parameters.

* Updated the `getServerSideProps` function in multiple pages
(`account/index.tsx`, `account/orders/[id].tsx`,
`account/orders/index.tsx`, `account/profile.tsx`,
`account/security.tsx`, `account/user-details.tsx`) to use
`getMyAccountRedirect` for handling redirection logic. This ensures
consistent behavior across all pages.

## How to test it

you can enable or disable the `enableFaststoreMyAccount` flag in
discovery.config.js and try to access any my account page.
`enableFaststoreMyAccount` true should use faststore
`enableFaststoreMyAccount` false should redirect to IO

### Starters Deploy Preview

TBD
