+++ packages/core/src/pages/account/index.tsx
@@ -1,6 +1,5 @@
-import storeConfig from 'discovery.config'
-
 import type { GetServerSideProps, NextPage } from 'next'
+import { getMyAccountRedirect } from 'src/utils/myAccountRedirect'
 
 const MyAccountRedirectPage: NextPage = () => {
   return null
@@ -9,26 +8,17 @@ const MyAccountRedirectPage: NextPage = () => {
 export const getServerSideProps: GetServerSideProps = async ({ query }) => {
   // TODO validate permissions here
 
-  if (storeConfig.experimental.enableFaststoreMyAccount) {
-    return {
-      redirect: {
-        destination: '/account/profile',
-        permanent: false,
-      },
-    }
-  }
+  const { isFaststoreMyAccountEnabled, redirect } = getMyAccountRedirect({
+    query,
+  })
 
-  const searchParams = new URLSearchParams()
-
-  for (const key in query) {
-    const value = query[key]
-    const values = Array.isArray(value) ? value : [value]
-    values.forEach((v) => v && searchParams.append(key, v))
+  if (!isFaststoreMyAccountEnabled) {
+    return { redirect }
   }
 
   return {
     redirect: {
-      destination: `${storeConfig.accountUrl}?${searchParams.toString()}`,
+      destination: '/account/profile',
       permanent: false,
     },
   }
